Trying differentiate punctuation from symbols
In https://github.com/ionide/ionide-fsgrammar/issues/144#issuecomment-543158400 I said that it was too hacky to maintains.

I creating this commit just so the work is not lost but IHMO we should not include it in the grammar.

PS: It doesn't cover all the cases and will probably never be possible to do it using regexes only